requirements.txt
astroid==1.6.2
attrs==17.4.0
bandit==1.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
coverage==4.5.1
decorator==4.3.0
dict-to-protobuf>=0.0.3.9
docker==3.3.0
docker-pycreds==0.2.3
Flask==1.0.2
get==0.0.39
gitdb2==2.0.3
GitPython==2.1.9
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.1.0
pbr==4.0.3
pluggy==0.6.0
post==0.0.26
protobuf==3.5.2.post1
public==0.0.65
py==1.5.3
pylint==1.8.3
pytest==3.5.0
pytest-cov==2.5.1
python-magic==0.4.15
PyYAML==3.12
query-string==0.0.28
ramrodbrain>=0.1.43
request==0.0.26
requests==2.18.4
rethinkdb==2.3.0.post6
six==1.11.0
smmap2==2.0.3
stevedore==1.28.0
urllib3==1.22
websocket-client==0.47.0
Werkzeug==0.14.1
wrapt==1.10.11
